Refactoring the UI: Moving tabs into separate class files. Some other changes too
